refactor unified to vertex ai
