+++ common/rest-vertex-ai-spring-api/README.md
