--- common/swagger/unified_swagger_codegen.json
@@ -1,7 +1,0 @@
-{
-  "basePackage":"ai.h2o.mojos.deploy.common.rest.unified",
-  "configPackage":"ai.h2o.mojos.deploy.common.rest.unified.config",
-  "modelPackage": "ai.h2o.mojos.deploy.common.rest.unified.model",
-  "apiPackage" : "ai.h2o.mojos.deploy.common.rest.unified.api",
-  "interfaceOnly" : true
-}
