+++ common/swagger/vertex-ai-swagger.yaml
