+++ gcp-vertex-ai-mojo-scorer/src/main/java/ai/h2o/mojos/deploy/gcp/vertex/ai/GcpVertexAiApplication.java
@@ -1,6 +1,6 @@
-package ai.h2o.mojos.deploy.gcp.unified;
+package ai.h2o.mojos.deploy.gcp.vertex.ai;
 
-import ai.h2o.mojos.deploy.gcp.unified.config.EnvironmentConfiguration;
+import ai.h2o.mojos.deploy.gcp.vertex.ai.config.EnvironmentConfiguration;
 import com.google.cloud.storage.Storage;
 import com.google.cloud.storage.StorageOptions;
 import org.springframework.boot.SpringApplication;
@@ -9,9 +9,9 @@ import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
 @EnableSwagger2
 @SpringBootApplication
-public class GcpUnifiedApplication {
+public class GcpVertexAiApplication {
   /**
-   * Wrapper application for running local rest scorer in Google AI Platform Unified.
+   * Wrapper application for running local rest scorer in Google Vertex AI.
    * Downloads pipeline.mojo and license.sig files from GCS before starting rest server.
    *
    * @param args N/A, application only requires environment variables
@@ -19,6 +19,6 @@ public class GcpUnifiedApplication {
   public static void main(String[] args) {
     Storage storage = StorageOptions.getDefaultInstance().getService();
     new EnvironmentConfiguration(storage).configureScoringEnvironment();
-    new SpringApplication(GcpUnifiedApplication.class).run(args);
+    new SpringApplication(GcpVertexAiApplication.class).run(args);
   }
 }
