+++ settings.gradle
@@ -7,10 +7,10 @@ include 'common:rest-jdbc-spring-api'
 include 'common:rest-spring-api'
 include 'common:transform'
 include 'common:kdb-java'
-include 'common:rest-unified-spring-api'
+include 'common:rest-vertex-ai-spring-api'
 include 'local-rest-scorer'
 include 'kdb-mojo-scorer'
 include 'aws-sagemaker-hosted-scorer'
 include 'gcp-cloud-run'
 include 'sql-jdbc-scorer'
-include 'gcp-unified-mojo-scorer'
+include 'gcp-vertex-ai-mojo-scorer'
